Add conditions if file does not exists.

+++ build/executions.sh
@@ -10,14 +10,18 @@ fi
 #
 # Assuming that the number of users is less than 1000.
 users=$(getUsers)
-echo $users | jq '.' > $export_path/users.json
+if [[ ! -f $exported_path/users.json ]]; then
+  echo $users | jq '.' > $export_path/users.json
+fi
 
 # Get user IDs.
 user_ids=$(echo $users | jq '.items[] | .id')
 
 # Get user info.
 for user in $user_ids; do
-  getUserInfo $user > $export_path/user.$user.json
+  if [[ ! -f $exported_path/user.$user.json ]]; then
+    getUserInfo $user > $export_path/user.$user.json
+  fi
 done
 
 # Get user chat history.
@@ -26,29 +30,30 @@ done
 # Define start index.
 start_index=0
 
-user_ids="3454640 1343698"
-
 for user in $user_ids; do
   count=1
 
-  while [[ $count -ne 0 ]]; do
-    chat_history=$(getUserChatHistory $user $start_index)
+  if [[ ! -f $export_path/user.$user.chat.$start_index.json ]]; then
+    while [[ $count -ne 0 ]]; do
+      chat_history=$(getUserChatHistory $user $start_index)
 
-    # Check whether to while loop again.
-    if [[ $(echo $chat_history | jq '.items | length') -eq 0 ]]; then
-      # Don't while loop again.
-      count=0
+      # Check whether to while loop again.
+      if [[ $(echo $chat_history | jq '.items | length') -eq 0 ]]; then
+        # Don't while loop again.
+        count=0
 
-      # Reset start index.
-      start_index=0
-    else
-      # Write to file.
-      echo $chat_history | jq '.' > $export_path/user.$user.chat.$start_index.json
+        # Reset start index.
+        start_index=0
+      else
+        # Write to file.
+          echo $chat_history | jq '.' > $export_path/user.$user.chat.$start_index.json
+        fi
 
-      # Increment start index by 1k.
-      start_index=$[$start_index+1000]
-    fi
-  done
+        # Increment start index by 1k.
+        start_index=$[$start_index+1000]
+      fi
+    done
+  fi
 
   # Reset start index.
   start_index=0
